Implement app enumeration on Linux platforms.
We can leverage the freedesktop application files to generate a list
of installed applications on the user's device, and even associate
icons to them as well.

Some issues of note:
 - This won't catch programs launched from a command line, or services
   started in the background (eg: systemd).
 - Qt doesn't seem to do as good of a job locating Icons as Gnome. In
   particular it only seems to find icons matching the current theme
   and has no fallback behavior.

As an alternate approach, we could interrogate the window manager for
a list of GUI applications that are running, but this presents other
problems that it A) becomes difficult to retain the user's selection
across restarts, and B) becomes impossible have applications contained
within the VPN at launch.