Add systemd service file and kill old versions on upgrade
When launching the D-Bus service as a bus-activatable service we
can start the daemon but we don't have a good way to stop it
during a software update. Adding a systemd service file allows
the OS to manage that service when interracting with the package
manager.

However, when upgrading from 2.2.0 and earlier the systemd service
file doesn't exist. For these cases we need a preinst hook to kill
off any old versions that might be running.

See: mozilla-mobile/mozilla-vpn-client#977