Install systemd service file via make install
In a recent commit, 7f44dc8, which added a systemd service file, we only
added it to the Debian package. This means that a user trying to install
from sources can miss the installation of this file via `make install`.
Leaving them with an incomplete installation.

See also: https://github.com/mozilla-mobile/mozilla-vpn-client/issues/1028