Move revision into upstream version.
It turns out that you can't change an orig.tar.gz once it has been
uploaded to a PPA unless you bump its upstream version (which makes
sense when you're dealing with truly upstream sources). However,
this means we have to change the upstream version on each build for
this to work.

I am not sure I like this approach yet. Consider this to be
experimental.