Fix potential routing loop during server switch
During a server switch, there's a period of time when an intermediate
hop can get removed but the final hop is still routing through it. This
can lead to a situation where the routing table forwards the packet back
into the wireguard interface but the destination doesn't match any peer,
and causes the packet to start again at hopindex 0.

To make this harder to occur, reorder the server switch sequence:
 1) Add the new peer.
 2) Add the new peer's routes.
 3) Remove the old peer's routes.
 4) Remove the old peer.