Multihop Tunnelling (#1376)
* First attempt at multihop tunneling for Linux

* Refactor multihop to use one interface with multiple peers

* Implement multihop for Windows (needs tunnel.dll fix)

As a note of caution, this requires a custom build of tunnel.dll with
fixes to disable the default route monitor and a fix for a deadlock in
the wireguard-go module. These changes don't have a home (yet).

* Import wireguard tunnel.dll go sources

* Remove default route monitor from tunnel.dll

* Remove hopindex from deactivate()

If the client is restarted without the daemon, the client can loose
track of how many hops are running, which causes problems when it
comes to deactivate the interface. We refactor the deactivate() call
so that it always deactivates all hops.

* Fix throughput stats when there are multiple peers

* Fix potential routing loop during server switch

During a server switch, there's a period of time when an intermediate
hop can get removed but the final hop is still routing through it. This
can lead to a situation where the routing table forwards the packet back
into the wireguard interface but the destination doesn't match any peer,
and causes the packet to start again at hopindex 0.

To make this harder to occur, reorder the server switch sequence:
 1) Add the new peer.
 2) Add the new peer's routes.
 3) Remove the old peer's routes.
 4) Remove the old peer.

* Linter fixes

* Fix windows multihop after split tunnel rebase

* Fix MacOS compile errors

* Move MacOS tunnel startup into wgutils

* Move MacOS peer and status management into wgutils

* Implement IPUtils for MacOS

* Fix reliability issues in MacOS tunnel startup

* Fix garbage interface flags set during ifup.

* Add MacosRouteMonitor class to manage the routing table

* Program routes via a routing socket

* Make wgutils mandatory, add server switching and multihop to MacOS

* Remove calls to logger.log(), fix merge cruft

* Remove one more logger.log() call from MacOS

* Fix compile error on Windows

* Fix invalid IPv6 DNS server in wg-quick config

* Enable multihop in dummy builds

* Fix MacOS DNS by implementing DnsUtilsMacos

* Remove macos helper.sh script

* Comments and copyright for multihop