MacOS: Fix squashed/overlarge status bar icon (#1823)
Fix a MacOS issue where the status bar icon appears horizontally squashed under the following conditions:
- Using MacOS Big Sur
- Using an external display
- The display has a device pixel ratio greater than 1

The icon also appeared much larger than other status bar icons, which is also fixed.

The underlying reason for this is [a bug in Qt 5.15](https://bugreports.qt.io/browse/QTBUG-88600) that causes the system tray icon class to not properly respect the device pixel ratio of the display. The irritating thing is that this bug has actually been fixed, but was released as part of Qt 5.15.3 which apparently is a [commercial-only patch release](https://www.qt.io/blog/commercial-lts-qt-5.15.3-released). This means we need to resort to an ugly hack, because capitalism.

The hack implementation (borrowed from the original bug report) uses an Objective-C technique called "swizzling" that makes it possible to replace the underlying implementation of an interface at runtime. The fix replaces the `setImage` method on `NSStatusBarButton` with a new version that correctly scales the image according to MacOS UI specifications.