Bound properties before onComplete
For Qt to set up the correct behaviors for the window buttons, the max size properties have to be set before the window is created.  They were previously set in onComplete.  This not only disables the minimize button, but corrects the double click behavior on the title bar, and makes the system menu items have the correct state.