Banner Fixes
The "No Internet" banner was fixed by ignoring NoConnectionError in MozillaVPN::errorHandle when we are in the State::StateConfiming state.
The intermittent "Unable To Connects" were caused by the periodic tasks that are scheduled.  I removed error reporting from the two tasks that did report errors on the periodic cycle.  The do not know the state of the app, so it is hard to accurately report from them.  We can look at cleaning some of the overall state up in another issue perhaps.