Banner Fixes (#2152)
* Banner Fixes

The intermittent "Unable To Connects/No internet" were caused by the periodic tasks that are scheduled.  The TaskControllerAction had a timeout timer that ran after 3 seconds.  This allowed the periodic tasks to run before the connection is properly established.  I changed the timeout logic in the TaskControllerAction to check the state every 3 seconds instead.  The first pass will always succeed.  From the 2nd onward if the state hasn't changed in the timer duration, then the task completes.  At that point the other tasks fire and the connection status is updated.

Added a cancel method to Task.  When we clear the task queue on connect, the tasks were still getting signal callbacks after execution.  The cancel ignores them if cancelled.  The cancel is a built in semantic now that can be overridden to do custom cancel logic such as signal disconnects, etc.