Update windowspingsender.cpp (#2195)
* Update windowspingsender.cpp

I was never able to reproduce this.  I think that is because the fixes I put in that have TaskControllerAction using a smarter watchdog vs. simple timer are mitigating it.  That said I think these errors should have resulted in the criticalPingError signal to be raised.  I've also added extra logging if the problem persists to make sure the address actually isn't invalid as the error states.

* Update macospingsender.cpp

Same fix as windows for when the ping can't go out.