Gracefully support QSessionManager termination (#2383)
When terminating the desktop session, the default termination provided by Qt often results in a hung logout session due to the minimize-on-quit behavior of the main window.

Furthermore, logging out of the desktop session with the VPN connected may leave the tunnel active even though the GUI has exited. This can lead to confusing connection behavior when logging back in. To remedy this, it is safer to disconnect the VPN when handling session termination.