Openssl script update (#2707)
* Update openssl_compile.bat

Updated to use a secure path for open ssl.  Script now requires admin to run (it is checked).

* Update openssl_compile.bat

Formatting

* Update openssl_compile.bat

Need to use the BUILDDIR variable.

* Update openssl_compile.bat

Added the -DOPENSSL_NO_ENGINES compile switch to disable engines, since we don't use them.

* Update openssl_compile.bat

Added fixed paths to the openssl dir in secured directories.  Added back no-engine, as that is actually needed for the config portion of the compile.