Tunnel log entries were not decoded correctly
On windows, the conversion from the fixed length buffer read from the tunnel's ringbuffer resulted in a 512b fixed length string being written to the log.  When encoded and sent to guardian for a support request that turned in to 512 /u0000s which caused it to be rejected by FxA for length.
I now truncate the buffer before conversion.  Also fixed a casting issue that was causing a warning.