Fix a crash in the CI - VPN-1989 (#3170)
The tasks are fully disconnected when they complete their job or when they are
deleted. Because of this, the ConnectionBenchmark object does not receie the
QObject::destroyed signal, it does not cleanup the list of tasks, and
eventually, it crashes.

This PR fixes this using a 3rd QObject (a sentinel). It also remove the
multiple connections using a `initialize` method.