Fix handling of controller state changes while benchmarking connection (#3163)
* fix: Handle connection info controller state

* fix: Return on connection stability change

* Fix a crash in CI - nullify the request when completed - VPN-1930 (#3087)

* fix: Handle changing controller state

* Fix a crash in the CI - VPN-1989

The tasks are fully disconnected when they complete their job or when they are
deleted. Because of this, the ConnectionBenchmark object does not receie the
QObject::destroyed signal, it does not cleanup the list of tasks, and
eventually, it crashes.

This PR fixes this using a 3rd QObject (a sentinel). It also remove the
multiple connections using a `initialize` method.

* chore: Clean up CI toggle

* chore: Simplify if-else condition

Co-authored-by: Andrea Marchesini <amarchesini@mozilla.com>