Start a server verification task on handshake failure.
When the user encounters a server unavailable notification error, this
often results in a support ticket. In order to provide better logs in
this case, we can attempt to ping the failed server and log what we
get back. Hopefully, this can help us identify if there is a genuine
reachability issue, or if this is some kind of UDP/firewall block.