Move version string into constants.cpp
Because the BUILD_ID is re-generated every time CMake reconfigures the
project, and this changes the command line macros, we find ourselves
forced to rebuild everything whenever we touch a makefile. This quickly
gets frustrating. So, to save my sanity, by moving the version string
into constants.cpp we only need to re-link the project on each change.