ci(taskcluster): make sure Android build failures turn the task red (… (#3461)
* ci(taskcluster): make sure Android build failures turn the task red (#3413)

We noticed that sometimes the Android builds fail but the task is still
green. Partly this is taskcluster/taskcluster#4590 as the task should
fail if one of its declared artifacts doesn't exist.

But partly the build script itself should fail the task if one of its
commands returns non-zero. This PR fixes that by enabling `set -e` in
all the build scripts.

* ci: rename 'where' -> 'which' in wasm builds