ci(signing): switch Android builds back to an AWS pool
The GCP build images we were using have an outdated Chain of Trust key
baked in and will need to be re-built with the new ones.

This means we can use these images for any release builds as they'll
fail the CoT verification when the signing tasks run.

For now let's just switch back to an AWS pool.