Fix split tunnelling with cgroupsv2
Unfortunately, it seems that we can't quite get the cgroup matching to
work with only a single pass, and instead we need to follow the same
approach as cgroupsv1 by adding separate mangle and NAT rules.

It's as though once a packet is marked, we are unable to change that
mark with a rule later in the chain, which kind of forces us to
duplicate the cgroup matching rule when deciding if NAT is necessary.