BUGFIX: Fix macos cmake config file
My experience with CMake is none, so I am not really sure what was going
on here. What I know is that I was getting the following error:

TARGET_BUNDLE_DIR is allowed only for Bundle targets

And with the change in this commit I don't get it anymore. Added this in
a separate commit so that it can be easily dropped if this was just me
doing something wrong.