Guarantee cmake overwrites CMAKE_OSX_DEPLOYMENT_TARGET variable
I am not completely sure why, but with caching this variable was not set
which prevented the Balrog enabling step and icon generation steps to
run for the MacOS build.

I attempted to move the `set` call for this variable to before the
`project` call as advised in
https://cmake.org/cmake/help/latest/variable/CMAKE_OSX_DEPLOYMENT_TARGET.html,
but that did not work.