BUGFIX: Disconnect all signal handlers on destruction of ws handler
Actually the only one giving me a headache was the onClose listener of
the internal websocket. QWebSocket is closed on destruction, that was
sometimes triggering `onClose` after destruction of the ws handler and
causing a segmentation fault error on tests.

For good measure I disconnected all signal handlers, why not?