VPN-2384 - Fix intermittent MacOS failure on websocket tests (#4153)
* BUGFIX: Disconnect all signal handlers on destruction of ws handler

Actually the only one giving me a headache was the onClose listener of
the internal websocket. QWebSocket is closed on destruction, that was
sometimes triggering `onClose` after destruction of the ws handler and
causing a segmentation fault error on tests.

For good measure I disconnected all signal handlers, why not?

* Simplify and follow best practices on web socket handler tests

I merged two tests that were mostly doing the same thing and
stopped relying on timeouts and simpy used QTRY_* APIs.