ci: move 'bundle-addons' task to the build kind
It turns out that Chain of Trust only supports a hardcoded list of kinds
that are allowed to be in the release pipeline. This was causing an
error as the `addons` kind was not known. We can get around this without
deploying a new scriptworker release by moving this task to the `build`
kind.

Jira: RELENG-968