Create a task to sign the addons-manifest (#4403)
* ci: fix 'release_artifacts' transforms when attributes not initialized

* ci: move 'bundle-addons' task to the build kind

It turns out that Chain of Trust only supports a hardcoded list of kinds
that are allowed to be in the release pipeline. This was causing an
error as the `addons` kind was not known. We can get around this without
deploying a new scriptworker release by moving this task to the `build`
kind.

* ci: ensure 'addons-bundle' task exposes files as direct artifacts

Previously we were only uploading a zip file, but the signing and
beetmover tasks that depend on this will need them to be directly
exposed as artifacts to work (since they just pass the artifact URLs
onto scriptworker without downloading them).

* ci: create a new task to sign the addons manifest

This task signs the addons-manifest using the `autograph_rsa` strategy.
The original file along with a detached signature will be uploaded as
artifacts.

Upon decrypting the detached signature, the `sha256` hash of the
original file will be obtained.

Jira: RELENG-968