Do not set Glean debug view tag by default
The debug view tag is a feature that should really only be used when
developing / testing Glean instrumentation, because it will redirect
pings to the Glean ping debug viewer (https://debug-ping-preview.firebaseapp.com/pings/MozillaVPN).

By leaving it default on we are in practice just flooding the data
pipeline and debug viewer for no reason.