Always enqueue a remove device task regardless if device is in DeviceModel
TaskRemoveDevice already deals with the possibility of the device
not being registered to the given user. No need to handle
that in multiple places.

Also by not launching task remove device, operations to take
user out of device limit state don't happen and that leaves users
trapped in device limit screen.