VPN-2934 - Fix device limit issues when logging out manually from other device (#4623)
* Start periodic operations timer whenver a token is available

1. That is all we need to run those operations
2. That ensures the operations are run also when in the device limit
   screen.

* Skip device check when in StateDeviceLimit

* Always enqueue a remove device task regardless if device is in
DeviceModel

TaskRemoveDevice already deals with the possibility of the device
not being registered to the given user. No need to handle
that in multiple places.

Also by not launching task remove device, operations to take
user out of device limit state don't happen and that leaves users
trapped in device limit screen.