Do not treat settings reset as a location change - VPN-3293
Note that this could be reproducible via a simple logout.  The bug is in the
retranslate request (ServerData::changed signal) when the language code changes
because of a setting reset.  The signal is received by the RecentConnection
model, which adds the "new" server location to its lists.

This PR renames VPNRecentConnections to VPNRecentConnectionsModel to avoid
conflicts with a nebula QML component.