Fix crash on linux in the pingsenders - VPN-3572
In a few ping sender implementations, we emit signals in loop (while-stmt and
for-stmt), but the slots connect to them, trigger the deletion of the entire object.
At the next cycle of the same loop, the object is not valid anymore and we crash.

The solution is to queue the slots and allow the ping-sender loops to complete
before processing the data.