Remove Glean.js from iOS
There is a name clash between the deprecated Glean module
and the Glean iOS SDK. I think the best way forward here is to just turn
Glean.js off on iOS.

I think this makes sense, because the data coming from iOS with Glean
is so weird, the validation would be pointless i.e. it would be hard
to validate against data is already very hard to reason about.