Put CFBundleIcons back into the plist template
I relent, brizental was right after all. The issue seems to be that Xcode
cloud is attempting to run a verification check on the plist that's in the
xcodeproj rather than the plist that's in the application bundle. This means
that even though we patch the plist file to include the app icons correctly,
Xcode clould still thinks that we are missing the icons.