Add Glean to the network extension target
Glean is normally added as a swift package, however that is
not supported by CMake yet and the workaround I found
(https://github.com/compnerd/swift-cmake-examples/tree/main)
expects the glean-swift package repository to include cmake
configuration files, which it doesn't.

I attempted to include iosglean as a dependency, but that cause a bunch
of issues. Finally I decided to just create a separate Glean target just
for the WebExt.'