Restore qtglean/Cargo.lock
Cargo.lock files are required so that a given version of Mozilla VPN
can be rebuilt from the same code.

The NixOS package cannot be built without Cargo.lock, because NixOS
enforces this kind of determinism using a sandbox that disables
network access for builds that don’t have a fixed output hash.

This undoes #6278.  See also #3341, #5236.

Signed-off-by: Anders Kaseorg <andersk@mit.edu>