Bug 1774273 - Part (1/2) - Add signingscript macos notarization task
MacOS notarization is moving from iscript to signingscript and requires a separate task.

There are no ways of testing notarization outside production. So, this change only adds the notarization task, but DOES NOT affect any other tasks.
The current signing still notarizes, and is treated as source/upstream by the beetmover task