Record health metrics on the Android daemon
The connection health checks on the Daemon and different from the
connection health checks on the client. This is a caveat worth noting at
analysis time. I have added a longer description to the metrics to make
this more explicit.

For this work I have considered that an unstable event happens when the
connection is deemed worh a silent server switch while a no signal even
happens when we have given up hope. Does that make sense?