Downgrade com.android.billingclient:billing-ktx to 5.2.0 (#7214) (#7219)
This was causing the crash described on VPN-5019. I don't really
understand why this issue. We are using a version of desugar_jdk_libs
that contains the aledgedly missing API.

I tried to add the desugar_jdk_libs with the right version as dependecy
to the vpnClient project, but then I get the error "Execution failed for
DexingNoClasspathTransform".

We may want to file a ticket to upgrad this back, but I guess unblocking
QA ASAP is more important for now.