ci: Use upstream version of `release_artifacts` transforms (#7276)
* ci: Add 'mozilla-taskgraph' to Decision task dependencies

* ci: Use upstream 'release_artifacts' transforms

Releng is undertaking an effort to consolidate much of the release logic
that gets copy/pasted from one project to the next. To that end, the
`upstream_artifacts` transforms have been moved to a new
`mozilla-taskgraph` package.

This pull requests adds a dependency on `mozilla-taskgraph 0.1.0` and
removes the `upstream_artifacts` implementation here.

The generate graphs are identical before / after this change.