VPN-4921 - Get all iOS logs to show up on console and log file (#7265)
* Refactor iOS log handler and move to Swift

Moving to Swift is not 100% necessary, I suppose. But to he honest I
only realized Objective-C was doing a very similar thing after I was
almost done with the move. Having less Objective-C in the codebase feels
like an improvement, so leaving it in.

* Log ALL iOS logs through IOSLogger

This is convenient for debugging with the Console.
Previously the VPN logs did not make it to the console.

* Address review comments