Close the DNSPingsender UDP socket on idle
The fluctuations of the VPN tunnel and the lifecycle
of the application itself seem to affect the UDP socket.

In iOS, the socket will deterministically go into an error
state when 1. The VPN is turned on 2. The device goes to sleep
3. The device is woken up 4. The VPN is toggled off. This situation
is addressed by a restart of the UDP socket.

Just starting and stopping it when necessary seems like a more
appropriate solution overall, the UDP socket doesn't need to be
alive while we are not sending pings anyways.