Revert MacOS universal builds
We have recently found out that something in between Golang and Qt is
resulting in a crash when compiled for aarch64/arm64 on MacOS. This crash
only occurs when attempting to verify installer signatures fetched from
Balrog, as that is the only place that MacOS uses Golang. As an emergency
workaround, we have decided to evaluate reverting to an x86_64 build and
relying on Rosetta for M1/M2 Macs.