Bring back force_heartbeat_failure inspector cmd
This is still needed to for functional tests. We need this to continue testing that even though a heartbeat failure can occur, it won't necessarily lead to the user seeing the something went wrong screen. The new command force_backend_failure bypasses some logic and forcibly shows the something went wrong screen so it can't be used in the testHeartbeat.js functional tests