Isolate the most used modules of the codebase into targets
These are the minimal semantic modules I had to separate
in order to be able to create an isolated target to test
the daemon stuff.

I could have gone more minimal here, yes. However that would
mean creating targets that don't necessarily make sense as a
group of functionalities bundled together.

Why do this at all?

Currently, if we want to create an isolated test suite for a
given module of this codebase we either have to re-add the
exact set of files our isolated stuff will require to work
and also all the dependencies these files have or... we have
to import all of the shared-sources in the very least
and also set up all the dependencies these sources have.

The approach introduced in this commit, creates isolated
buildable targets that can be linked to other isolated
targets to create minimal buildable executables that only
have the strictly necessary stuff for the bundle to build.

This was done, as will become clear in the next commits, in
order to create an executable for the daemon tests. The daemon
includes leakdetector and logging. Most of the other stuff
I modularized here was because of the dependencies of logging.