Fix testSettings.js functional test on MacOS
The tests appear to be failing on MacOS due to the DNS/privacy modals
taking time to make their exit transitions, during which the next test
step winds up clicking the modal instead of the intended QML items. To
fix the tests, we need to wait for the modal loaders to become inactive
before proceeding.