Move version number definition to 'version.txt'
Our release automation (e.g shipit and treescript) assumes that the
version number is defined in a file called 'version.txt'. While
supporting the current version definition in CMakeLists.txt should be
possible, it will be much simpler to pull it out to the standalone file.
This way we can also get rid of that regex based version parser script.

Bug: 1871138