Checkpoint 1: Create the concept of a CMake module and apply it to settings/
This is the first checkpoint as we can't move ahead from here without
turning other parts of the codebase into modules OR linking our module
to the whole app. Given we are YOLOing at the moment, we will take the
hardest route: make other parts of the codebase into modules. Linking
to the module to the whole app as an intermediary step is z quick fix,
but it removes one big benefit of the module system which is super
quick build times for individual modules.