[VPN-6080] Refactor beetmover tasks to be distinct to each shipping phase (#8931)
* ci: Refactor some beetmover logic out of the 'tasks' hot loop

* ci: Add version to beetmover ship-client task's description

* ci: Remove 'if-dependencies', 'worker' and 'treeherder' keys from beetmover tasks

The `beetmover` transforms weren't using these keys in any way, nor
were they forwarding these keys on to the next set of transforms. So
the configs defined here were no-ops.

* ci: Consolidate beetmover logic to 'task-defaults'

* ci: Add a schema for beetmover tasks

* [VPN-6080] create distinct beetmover tasks for each shipping phase

This refactors the beetmover transform to generate distinct tasks for
each shipping phase. This is more inline with how other projects work
and is a pre-requisite for integrating VPN with Shipit, which has
assumptions around how the tasks are set up.

Issue: #8919