Fix factory reset functional tests
This is a fun one! The 'go back' button winds up hiding underneath the navbar,
So when we open the factory reset screen and try to click it, we go to the messages
view instead and the test gets stuck. We need to add a step to explicitly scroll to
the bottom of the view before clicking.