Set objectName in MZInAppAuthenticationBase to make it scrollable
This fixes a a broken authentication test where the cancel button winds
up off the bottom of the screen and we are unable to scroll the view
because there is no objectName on MZInAppAuthenticationBase.

However, this change is complicated by the maze of objectName
inheritance in the in-app-authentication QML elements.