Fix split tunnelling after firewall changes
When the firewall default policy is DROP, it becomes difficult to
know when to accept or drop inbound packets that relate to split
tunneling, as the inbound packets are likely to violate the firewall
policy. In order to do this, we switch from using a ctzone for
excluded traffic to keeping everything in one zone, but instead using
a ctmark. This allows us to differentiate inbound packets that are
headed to an excluded application.

Turns out we had a few bugs in here too. For one, setting a ctzone or
ctmark before the `conntrack` priority is ignored as conntrack resets it
to zero, and we also need to move the cgroup matching rule to an earlier
priority