Stabilize tests which depend on telemetry
A common test instability seems to be from clicking on a thing, then
checking that the telemetry has been received. But the processes by
which we gather telemetry don't always occur on click. A better
pattern in many of these cases is to click and then wait for pings
to be recorded.

We add a helper: vpn.waitForGleanValue() to do just this.