Another attempt to fix test races in ViewGetHelp
I think there is a race condition here in that we can open the GetHelp
view and then click the back button while the view is still loading. We
potentialy wind up clicking the back button in before Component.onCompleted
is called and causing the MZNavigator and stackviews to get out of wack.