Nit: Remove GHA retries for failed tests (#9504)
* Remove GHA retries for failed tests

These just seem to make the tests take far longer than necessary and
make it harder to spot flaky tests. Besides, we already have the
"--retries 3" option passed to mocha when running tests. We don't
need retries ontop of our retries.

* Stabilize tests which depend on telemetry

A common test instability seems to be from clicking on a thing, then
checking that the telemetry has been received. But the processes by
which we gather telemetry don't always occur on click. A better
pattern in many of these cases is to click and then wait for pings
to be recorded.

We add a helper: vpn.waitForGleanValue() to do just this.

* Fix testAuthenticationInAppAccountCreationNav race when loading getHelp

* Maybe wait for getHelpStackView to be ready instead

* Add logging of inspector command errors.

* Fix stackview test race a better way: define a new 'ready' property

The root problem is that the stackView 'ready' query only checks for
the stackView being not-busy, but this misses the case where nothing
has been loaded yet. Let's define a new property 'ready' that does
more or less the same thing, but also checks for an empty stackview.

* Give the webdriver more time to startup