Add routing table setup back for Linux.
Once the default-route-suppression rule was removed, it turns out that
we really do need to setup the wireguard routing table to exclude LAN
addresses. This can easily be done using the RTN_THROW route type.

But then once the throw routes are in place, we can't actually reach
the internal Mulvad address space, so we need to add unicast routes
back too.

There is a bunch of duplicated work between the throw and unicast
routes though. So it still results in an ugly looking routing table.