VPN-5992 Allow extensions to request Access to the Proxy Network (#9639)
This pull request introduces support for a new StateOnPartial state in the controller, enhancing the VPN's state management. It also adds an ActivationPrincipal enum to track whether a connection was initiated by a user or an extension, along with corresponding permission checks for deactivation. Additionally, this update includes several new methods and refines existing ones to handle the new state and activation principal, ensuring improved VPN control and security.


* WORKS only tests missing

* Add tests

* Remove backdoor >:(

* Todo: get a spellchecker for vscode

* Skip StateConnecting for the Extension

* Bonus: Remove unused State, making this PR state Neutral

* Lint

* Update tests/functional/helper.js

Co-authored-by: Lesley Norton <lesley@mozilla.com>

* Update src/controller.h

Co-authored-by: Lesley Norton <lesley@mozilla.com>

* revert oops

* Fix functionalTests

* ugh

---------

Co-authored-by: Lesley Norton <lesley@mozilla.com>