Add route capturing to WindowsRouteMonitor
This attempts to make up for the lack of policy-based routing on Windows
by interactively updating the routing table to force non-local traffic
into the VPN tunnel. The gist of the algorithm being deployed here is to
dump the routing table, and then dynamically duplicate any routes which
might try to leak traffic but with a lower metric to force it back into
the VPN anyways.