Migrate to XdgStartAtBootWatcher too
For host applications (eg: not in a Flatpak/Snap sandbox), there is some
guesswork applied in the XDG Portals to figure out the application ID.
If that guess is wrong, then the secrets API will give us a different
key because it thinks we are a different application.

Unfortunately, this guess goes very wrong when launched via the old
/etc/xdg/autostart handlers, which results in the keys getting wiped
upon boot. Luckily for us, just switching to the XDG Background
portal fixes it for us.