Bug 1927890 - Hardcode beetmover-apt-promote task to level 1 worker (#9998) (#9999)
For now there isn't a good place for us to import the candidate package
during the promote phase for QA to be able to test. So we use the
"staging" apt repo. However, this lives in the non production GCP
project, so we have to use the level 1 workers.

In the future we'll want to either:
1. Create a second apt repo in the production GCP project that we can
   use for QA prior to shipping.
2. Import the package into the main `mozillavpn` production apt repo but
   with a different name (e.g include `rc###` or similar). Then add a
   repackage task that renames it to the proper name during the ship
   phase.

Co-authored-by: Andrew Halberstadt <ahal@mozilla.com>