Linux: Drop CAP_NET_RAW capability from binary (#10016)
I have been fighting with the XdgCryptoSettings on my Fedora machine
which fails to retrieve the secret. The XDG Desktop portal throws an
error because the clients procfs entry is root-owned. It turns out this
is because the process image has elevated capabilities and we need to
drop them before XDG will let us read secrets.

So, why do we have CAP_NET_RAW granted at all? Well, it was added to
enable the client to send ICMP pings using a raw socket if the kernel
doesn't allow us to use ICMP sockets. However, we now have workarounds
that use TCP for connection health monitoring too, so it's not really
needed anymore.