Turn off Xcode's ENABLE_DEBUG_DYLIB option
This option breaks the Qt/iOS application setup. Qt tries to override the
program entrypoint on iOS, and directs it to _qt_main_wrapper() instead
using a linker argument, but these arguments get dropped when building as
a dylib, causing Xcode to skip _qt_main_wrapper() and invokes main()
directly.