Fix static assert in DeviceModel
We encounter an assert in Xcode 16 because the sortCallback method
used in the DeviceModel does not satisfy strict-weak ordering due
to the special case added for the current device. To fix this we
can simplify the callback so that it sorts by creation time and
move the current device as a second step.