Implement mock onboarding via the local socket controllers
When running on mobile, we need to support a speculative onboarding
connection in the controller. This is a special case handled in the
iOS and Android controllers.

To try and generalize this, lets add an onboarding state and reason
to the controller and pass it down to the local-socket based daemons
so that they can setup permissions and emit a disconnection to end
the onboarding.