Xcode 16 build fixes (#10148)
* Remove -Xlink=-force:multiple hack
* Add framework search paths for IOSGlean
* Turn off Xcode's ENABLE_DEBUG_DYLIB option

This option breaks the Qt/iOS application setup. Qt tries to override the
program entrypoint on iOS, and directs it to _qt_main_wrapper() instead
using a linker argument, but these arguments get dropped when building as
a dylib, causing Xcode to skip _qt_main_wrapper() and invokes main()
directly.

* Fix static assert in DeviceModel

We encounter an assert in Xcode 16 because the sortCallback method
used in the DeviceModel does not satisfy strict-weak ordering due
to the special case added for the current device. To fix this we
can simplify the callback so that it sorts by creation time and
move the current device as a second step.

* Bump conda golang to 1.22
* Update 3rdparty/wireguard-apple to fix Xcode16 compile fail